LPD-8518 Remove service from DLPortletToolbarContributorHelper, rename it to DLPortletToolbarContributorUtil pass dlAppLocalService as paramater and apply to usages
