+++ modules/apps/document-library/document-library-web/src/main/java/com/liferay/document/library/web/internal/portlet/toolbar/contributor/DLPortletToolbarContributor.java
@@ -6,9 +6,10 @@
 package com.liferay.document.library.web.internal.portlet.toolbar.contributor;
 
 import com.liferay.document.library.constants.DLPortletKeys;
+import com.liferay.document.library.kernel.service.DLAppLocalService;
 import com.liferay.document.library.portlet.toolbar.contributor.DLPortletToolbarContributorContext;
-import com.liferay.document.library.web.internal.portlet.toolbar.contributor.helper.DLPortletToolbarContributorHelper;
 import com.liferay.document.library.web.internal.portlet.toolbar.contributor.helper.MenuItemProvider;
+import com.liferay.document.library.web.internal.portlet.toolbar.contributor.util.DLPortletToolbarContributorUtil;
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerList;
 import com.liferay.osgi.service.tracker.collections.list.ServiceTrackerListFactory;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManager;
@@ -68,14 +69,14 @@ public class DLPortletToolbarContributor extends BasePortletToolbarContributor {
 			WebKeys.THEME_DISPLAY);
 
 		if (_isDLPortlet(themeDisplay) &&
-			!_dlPortletToolbarContributorHelper.isShowActionsEnabled(
+			!DLPortletToolbarContributorUtil.isShowActionsEnabled(
 				themeDisplay, portletRequest)) {
 
 			return null;
 		}
 
-		Folder folder = _dlPortletToolbarContributorHelper.getFolder(
-			themeDisplay, portletRequest);
+		Folder folder = DLPortletToolbarContributorUtil.getFolder(
+			_dlAppLocalService, themeDisplay, portletRequest);
 
 		List<MenuItem> menuItems = new ArrayList<>();
 
@@ -173,12 +174,11 @@ public class DLPortletToolbarContributor extends BasePortletToolbarContributor {
 		return false;
 	}
 
+	@Reference
+	private DLAppLocalService _dlAppLocalService;
+
 	private ServiceTrackerList<DLPortletToolbarContributorContext>
 		_dlPortletToolbarContributorContexts;
-
-	@Reference
-	private DLPortletToolbarContributorHelper
-		_dlPortletToolbarContributorHelper;
 
 	@Reference
 	private FeatureFlagManager _featureFlagManager;
